+++ app/controllers/app.js
@@ -262,10 +262,13 @@ function handleQuery(req, res) {
                 if ( cache_policy == 'persist' ) {
                   res.header('Cache-Control', 'public,max-age=31536000'); // 1 year
                 } else {
-                  // TODO: set ttl=0 when tableCache[sql_md5].may_write is true ?
-                  var ttl = 3600;
+                  // don't cache the result !
+                  // NOTE: both Varnish and Cloudfront ignore this header,
+                  //       X-Cache-Channel is for controlling Varnish
+                  //       Cloudfront can only be skipped perturbating the URL
+                  //       
                   res.header('Last-Modified', new Date().toUTCString());
-                  res.header('Cache-Control', 'no-cache,max-age='+ttl+',must-revalidate,public');
+                  res.header('Cache-Control', 'no-cache,max-age=0,must-revalidate,public');
                 }
 
                 return result;